+++ lib/website/View/web_loginage.dart
@@ -1,3 +1,4 @@
+import 'package:cmt_projekt/website/View/web_createaccountwidget.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/painting.dart';
@@ -33,98 +34,107 @@ class _WebLoginPageState extends State<WebLoginPage> {
               ),
             ],
           ),
-          Column(
-            children: [
-              Card(
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(40),
-                ),
-                elevation: 3,
-                child: Container(
-                  padding: const EdgeInsets.all(30),
-                  width: 400,
-                  height: 850,
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                    children: [
-                      Column(
-                        mainAxisAlignment: MainAxisAlignment.center,
-                        children: [
-                          const Text(
-                            'Välkommen',
-                            style: TextStyle(
-                                fontSize: 30, fontWeight: FontWeight.bold),
-                          ),
-                          const SizedBox(
-                            height: 40,
-                          ),
-                          TextFormField(
-                            decoration: const InputDecoration(
-                              labelText: 'E-post eller telefonnummer',
+          FittedBox(
+            child: Column(
+              children: [
+                Card(
+                  shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(40),
+                  ),
+                  elevation: 3,
+                  child: Container(
+                    //Här kan ni ta
+                    padding: const EdgeInsets.all(30),
+                    width: 500,
+                    height: MediaQuery.of(context).size.height,
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                      children: [
+                        Column(
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          children: [
+                            const Text(
+                              'Välkommen',
+                              style: TextStyle(
+                                  fontSize: 30, fontWeight: FontWeight.bold),
                             ),
-                          ),
-                          TextFormField(
-                            decoration: InputDecoration(
-                              labelText: 'Lösenord',
-                              suffixIcon: IconButton(
-                                hoverColor: Colors.transparent,
-                                splashRadius: null,
-                                splashColor: Colors.transparent,
-                                icon: Icon(
-                                  // Based on passwordVisible state choose the icon
-                                  _passwordVisible
-                                      ? Icons.visibility
-                                      : Icons.visibility_off,
-                                  color: Theme.of(context).primaryColorDark,
-                                ),
-                                onPressed: () {
-                                  // Update the state i.e. toogle the state of passwordVisible variable
-                                  setState(() {
-                                    _passwordVisible = !_passwordVisible;
-                                  });
-                                },
+                            const SizedBox(
+                              height: 40,
+                            ),
+                            TextFormField(
+                              decoration: const InputDecoration(
+                                labelText: 'E-post eller telefonnummer',
                               ),
                             ),
-                            obscureText: !_passwordVisible,
-                          ),
-                          Align(
-                            alignment: Alignment.bottomRight,
-                            child: TextButton(
+                            TextFormField(
+                              decoration: InputDecoration(
+                                labelText: 'Lösenord',
+                                suffixIcon: IconButton(
+                                  hoverColor: Colors.transparent,
+                                  splashRadius: null,
+                                  splashColor: Colors.transparent,
+                                  icon: Icon(
+                                    // Based on passwordVisible state choose the icon
+                                    _passwordVisible
+                                        ? Icons.visibility
+                                        : Icons.visibility_off,
+                                    color: Theme.of(context).primaryColorDark,
+                                  ),
+                                  onPressed: () {
+                                    // Update the state i.e. toogle the state of passwordVisible variable
+                                    setState(() {
+                                      _passwordVisible = !_passwordVisible;
+                                    });
+                                  },
+                                ),
+                              ),
+                              obscureText: !_passwordVisible,
+                            ),
+                            Align(
+                              alignment: Alignment.bottomRight,
+                              child: TextButton(
+                                onPressed: () {},
+                                child: const Text('Glömt lösenord?'),
+                              ),
+                            ),
+                            const SizedBox(
+                              height: 30,
+                            ),
+                            ElevatedButton(
                               onPressed: () {},
-                              child: const Text('Glömt lösenord?'),
+                              child: const Text('Logga in'),
                             ),
-                          ),
-                          const SizedBox(
-                            height: 30,
-                          ),
-                          ElevatedButton(
-                            onPressed: () {},
-                            child: const Text('Logga in'),
-                          ),
-                        ],
-                      ),
-                      Column(
-                        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                        children: [
-                          ElevatedButton(
-                            onPressed: () {},
-                            child: const Text(
-                              'Gästläge',
-                              style: TextStyle(fontStyle: FontStyle.italic),
+                          ],
+                        ),
+                        Column(
+                          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                          children: [
+                            ElevatedButton(
+                              onPressed: () {},
+                              child: const Text(
+                                'Gästläge',
+                                style: TextStyle(fontStyle: FontStyle.italic),
+                              ),
                             ),
-                          ),
-                          TextButton(
-                            onPressed: () {},
-                            child:
-                                const Text('Inget konto? Registrera dig här'),
-                          ),
-                        ],
-                      ),
-                    ],
+                            TextButton(
+                              onPressed: () {
+                                showDialog(
+                                    context: context,
+                                    builder: (context) {
+                                      return const WebCreateAccountWidget();
+                                    });
+                              },
+                              child:
+                                  const Text('Inget konto? Registrera dig här'),
+                            ),
+                          ],
+                        ),
+                      ],
+                    ),
                   ),
                 ),
-              ),
-            ],
+              ],
+            ),
           ),
         ],
       ),
